+++ controllers/story.controller.js
@@ -1,48 +1,57 @@
+require("dotenv").config();
 const db = require("../db/story.query");
 
 async function getStoryById(req, res) {
   const { storyId } = req.params;
-  try {
-    const results = await db.getStoryById(storyId);
-    const { rows } = results;
+  if (api_key && api_key === process.env.API_KEY) {
+    try {
+      const results = await db.getStoryById(storyId);
+      const { rows } = results;
 
-    if (rows.length === 0) {
-      return res.status(404).json({ error: "Story was not found" });
+      if (rows.length === 0) {
+        return res.status(404).json({ error: "Story was not found" });
+      }
+
+      return res.status(200).json(rows[0]);
+    } catch {
+      return res.status(500).json({ error: "Internal Server Error" });
     }
-
-    return res.status(200).json(rows[0]);
-  } catch {
-    return res.status(500).json({ error: "Internal Server Error" });
   }
 }
 
 async function updateStoryByID(req, res) {
   const { storyId } = req.params;
-  const data = {
-    title: req.body.title,
-    summary: req.body.summary,
-    type: req.body.type,
-    count: req.body.count,
-    year: req.body.year,
-    special: req.body.special,
-  };
-
-  try {
-    await db.updateStoryByID(storyId, data);
-
-    return res.status(201).json({ message: "Updated story successfully" });
-  } catch {
-    return res.status(500).json({ error: "Internal Server Error" });
+  const api_key = req.query.api_key ?? null;
+  if (api_key && api_key === process.env.API_KEY) {
+    const data = {
+      title: req.body.title,
+      summary: req.body.summary,
+      type: req.body.type,
+      count: req.body.count,
+      year: req.body.year,
+      special: req.body.special,
+    };
+    try {
+      await db.updateStoryByID(storyId, data);
+      return res.status(201).json({ message: "Updated story successfully" });
+    } catch {
+      return res.status(500).json({ error: "Internal Server Error" });
+    }
   }
 }
 
 async function deleteStoryByID(req, res) {
   const { storyId } = req.params;
-  try {
-    await db.deleteStoryByID(storyId);
-    return res.status(200).json({ message: "Story was deleted successfully" });
-  } catch {
-    return res.status(500).json({ error: "Internal Server Error" });
+  const api_key = req.query.api_key ?? null;
+  if (api_key && api_key === process.env.API_KEY) {
+    try {
+      await db.deleteStoryByID(storyId);
+      return res
+        .status(200)
+        .json({ message: "Story was deleted successfully" });
+    } catch {
+      return res.status(500).json({ error: "Internal Server Error" });
+    }
   }
 }
